scripts/pyproject.toml
[tool.poetry]
name = "tracking-sample"
version = "0.1.0"
description = ""
authors = ["chck <shimekiri.today@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mlflow = "^1.17.0"
google-api-python-client = "^2.9.0"
google-cloud-storage = "^1.38.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
